Have a basic app. Can start and stop tracking.
Have combined all three axes into one  to represent speed. have yet to
calculate noise. Not sure if this should be done per axis or on total
value for greater accuracy. There is a dummy value, and a method that
checks for wifi connection (yet to be connected to anything).