IT WORKS! FUCK YEAH!
Turns out I needed to add the INTERNET permission. Also networking
cannot be done on the main thread after Honeycomb, so it kept giving an
exception. It can now connect to the database and upload documents.
Networking is done in an async thread. The same should  be done for file
I/O.