does activity recognition, but not able to pass results back
activity recognition worked when made into a separate class and placed
as a service in the manifest. Having problems getting the information
back to the service calling it. Goes wrong when putExtra with a result
receiver is sent with the intent. not sure why. possibly something to do
with string value, even though I've given it a full package path name.
Also the activity isn't stopping when it should. at least it uses very
little power anyway. but need to look into this.