in the process of porting uploading to a service instead of async task
also added an utils class for shared constants and trying to figure out
what to do about uploading onReceive. it keeps throwing fits. finding a
file is DEFINITELY not a good idea of things to do in on receive.
onSensorChanged is doing too much work, and keep getting GC_FOR_ALLOC
when not still. try porting some of it out.