low pass filter and data stored in a custom linked list
a lot of changes back and forth have happened since the last commit.
tried storing data as arrayList, which almost seemed worse than string
concatenation, even with a large preallocated size. Implemented a strict
low-pass filter. It might need to be a little coarser, but gives nice
smooth curves for now. Need to actually activate the orientation
feature. The toString function of the linked list is REALLY slow and
happening on main thread, which is superbad. Particularly for storage
without wifi. You should really change this.